Build out rough component folder structure and fetch app data
Added folders for what I anticipate to be the necessary components for this module. Added basic markup to the TextAndVideoCard to get it working enough to import and connected app data to the component via props. Confirmed that data was flowing with console.logs which will be removed later.

Also installed sass for use later.